// src/modules/balancer-sdk/sor/type.ts
export interface GetSwapsInput {
  tokenIn: string;
  tokenOut: string;
  swapType: GqlSorSwapType;
  swapAmount: string;
  swapOptions: GqlSorSwapOptionsInput;
  tokens: PrismaToken[];
}

export interface SorConfig {
  chainId: number;
  vault: string;
  weth: string;
  staBal3Pool?: { id: string; address: string };
  usdcConnectingPool?: { id: string; usdc: string };
  wETHwstETH?: { id: string; address: string };
  lbpRaisingTokens?: string[];
}

export type NoNullableField<T> = {
  [P in keyof T]: NonNullable<T[P]>;
};

export enum SwapTypes {
  SwapExactIn,
  SwapExactOut,
}
// ...
export interface SwapV2 {
  poolId: string;
  assetInIndex: number;
  assetOutIndex: number;
  amount: string;
  userData: string;
}

export interface SwapInfo {
  tokenAddresses: string[];
  swaps: SwapV2[];
  swapAmount: BigNumber;
  swapAmountForSwaps: BigNumber; // Used with stETH/wstETH
  returnAmount: BigNumber;
  returnAmountFromSwaps: BigNumber; // Used with stETH/wstETH
  returnAmountConsideringFees: BigNumber;
  tokenIn: string;
  tokenInForSwaps?: string; // Used with stETH/wstETH
  tokenOut: string;
  tokenOutFromSwaps?: string; // Used with stETH/wstETH
  marketSp: string;
  routes: SwapInfoRoute[];
}

export interface SwapInfoRoute {
  tokenIn: string;
  tokenInAmount: string;
  tokenOut: string;
  tokenOutAmount: string;
  share: number;
  //hops in this route, properly ordered
  hops: SwapInfoRouteHop[];
}

export interface SwapInfoRouteHop {
  tokenIn: string;
  tokenInAmount: string;
  tokenOut: string;
  tokenOutAmount: string;
  poolId: string;
}